+++ pkg/proxy/oauth_proxy.go
@@ -1,0 +1,9 @@
+package proxy
+
+const (
+	GrantTypeAuthCode     = "authorization_code"
+	GrantTypeUserCreds    = "password"
+	GrantTypeRefreshToken = "refresh_token"
+	GrantTypeClientCreds  = "client_credentials"
+	GrantTypeUmaTicket    = "urn:ietf:params:oauth:grant-type:uma-ticket"
+)
