+++ pkg/testsuite.go/constant.go
@@ -1,0 +1,16 @@
+package testsuite
+
+const (
+	FakeAdminRole          = "role:admin"
+	FakeAdminRoleURL       = "/admin*"
+	FakeAuthAllURL         = "/auth_all/*"
+	FakeClientID           = "test"
+	FakeSecret             = "test"
+	FakeTestAdminRolesURL  = "/test_admin_roles"
+	FakeTestRole           = "role:test"
+	FakeTestRoleURL        = "/test_role"
+	FakeTestWhitelistedURL = "/auth_all/white_listed*"
+	TestProxyAccepted      = "Proxy-Accepted"
+	ValidUsername          = "test"
+	ValidPassword          = "test"
+)
